Commit HW_210326_2

+++ src/HW_210326/AirTransport.java
@@ -1,12 +1,12 @@
 package HW_210326;
 
 public abstract class AirTransport extends Transport {
-    protected short wingspan; //meter
-    protected short minimumRunwayLength; //meter
+    protected int wingspan; //meter
+    protected int minimumRunwayLength; //meter
 
     public AirTransport(int power, int maximumSpeed, int weight, String model, int wingspan, int minimumRunwayLength) {
         super(power, maximumSpeed, weight, model);
-        this.wingspan = (short) wingspan;
-        this.minimumRunwayLength = (short) minimumRunwayLength;
+        this.wingspan = wingspan;
+        this.minimumRunwayLength = minimumRunwayLength;
     }
 }
